+++ Artist-Recruitment-Website/Controllers/ReviewController.cs
@@ -1,4 +1,5 @@
-﻿using BL.Services.Interface;
+﻿using BL.Services;
+using BL.Services.Interface;
 using Microsoft.AspNetCore.Mvc;
 using Models;
 
@@ -13,10 +14,12 @@ namespace Artist_Recruitment_Website.Controllers
             _reviewService = reviewService;
         }
 
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(int id)
         {
-            var reviews = await _reviewService.GetAllReviewsAsync();
-            return View(reviews);
+            var items = await _reviewService.GetAllReviewsAsync();
+            var res = items.Where(op => op.ArtistProfileId == id);
+            ViewBag.ArtistProfileId = id;
+            return View(res);
         }
 
         public async Task<IActionResult> Details(int id)
@@ -28,9 +31,12 @@ namespace Artist_Recruitment_Website.Controllers
             return View(review);
         }
 
-        public IActionResult Create()
+        public IActionResult Create(int id)
         {
-            return View();
+            var review = new Review();
+            review.ArtistProfileId = id; // or set any other required properties
+            ViewBag.ArtistProfileId = id;
+            return View(review);
         }
 
         [HttpPost]
@@ -40,28 +46,8 @@ namespace Artist_Recruitment_Website.Controllers
                 return View(review);
 
             await _reviewService.AddReviewAsync(review);
-            return RedirectToAction(nameof(Index));
+            return RedirectToAction(nameof(Index), new { id = review.ArtistProfileId });
         }
-
-        public async Task<IActionResult> Edit(int id)
-        {
-            var review = await _reviewService.GetReviewByIdAsync(id);
-            if (review == null)
-                return NotFound();
-
-            return View(review);
-        }
-
-        [HttpPost]
-        public async Task<IActionResult> Edit(Review review)
-        {
-            if (!ModelState.IsValid)
-                return View(review);
-
-            await _reviewService.UpdateReviewAsync(review);
-            return RedirectToAction(nameof(Index));
-        }
-
         public async Task<IActionResult> Delete(int id)
         {
             var review = await _reviewService.GetReviewByIdAsync(id);
@@ -72,10 +58,17 @@ namespace Artist_Recruitment_Website.Controllers
         }
 
         [HttpPost, ActionName("Delete")]
+        [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            // Load the item first
+            var item = await _reviewService.GetReviewByIdAsync(id);
+            if (item == null)
+                return NotFound();
+
             await _reviewService.DeleteReviewAsync(id);
-            return RedirectToAction(nameof(Index));
+            // Now you can use item.ArtistProfileId for the redirect
+            return RedirectToAction(nameof(Index), new { id = item.ArtistProfileId });
         }
     }
 }
